[BREAKING] - Remove usage of StateEvent in Spark state events
Using "complex" proto schema with nested object is not really usefull
for our case and leads to more complex generic reader mechanism
So we decide to remove it.
Also take the oppoturnity of this breaking change to rename one field
end to move header event in a more appropriate proto file
This change will induce loss of some events during the bump.

Change-Id: Ic8fc76c171a62c653fe9fbb8391c26dd0fa8f396