Add process network statistics
Add processrx & processtx properties that show bytes received & sent by the
sockets associated to this process or forked one.
Based on command "ss -tinp" which we parse and resolve with the list of
processes to handle forked process and sum them up to the initial JVM process

Notes:
- Handle only IPv4
- based on the refresh rate, it can miss connections that may be short lived
between 2 collections. So this is only an estimation, not accurate numbers.