Add Presto module based on StandaloneModule
Presto use specific classloader for plugin with no parent.
This leads to garmadon class not being available for presto plugins
and also all libraries used by a plugin to not be loaded by parent classloader
so not being available for garmadon agent.
Using reflections on FileSystemTracer and ensuring that garmadon agent classes
are loaded from the App classloader and not the plugin one permits to use
trace hdfs call