[HDFS exporter] Allow setting Proto Parquet writer's "block size"
This is in fact the buffer size before flushing to disk. Setting it too high (128MB by default)
leads to OoM as entries are stored in-memory. Setting it too low would lead to constant writes.
Also had to set manually pageSize, taking the default value as set the previous constructor