[reader] serialize timestamp with along header and body
Currently ES reader is sending timestamp as a special field but hadoop
reader is using parquet and was not sending it.

This review adds a field `timestamp` when merging header and body to a
single protobuf message.

This field comes from GarmadonMessage.getTimestamp which is the
timestamp in millisecond of the event.