Enrich events related to application running on the cluster on ES reader
In order to be able to filter vcore/memory metrics per application_name, users, frameworks or component
we have to enrich those events with informations from AppEvent and events from containers (Doing joins on elasticsearch is not possible/efficient).
As component is only a notion from the framework we can only get this information from containers class parsing.
This information could also be set to UNKNOWN if we don't know the class which make impossible to display AM metrics or workers metrics.
In order to do so we are retreiving am container id from RM and compare it from current ccontainer id and update the component to APP_MASTER otr APP_SLAVE.