[HDFS EXPORTER] Commit pending files when revoking a partition, instead of dropping them
From ConsumerRebalanceListener' doc:
"It is guaranteed that all consumer processes will invoke onPartitionsRevoked prior to any
process invoking onPartitionsAssigned. So if offsets or other state is saved in the
onPartitionsRevoked call it is guaranteed to be saved by the time the process taking over
that partition has their onPartitionsAssigned callback called to load the state."

We used to drop because we (actually, I) thought the behavior was the other way around, i.e.
_assigned_ called before _revoked_