Add support for key value yarn.tags
It may be interesting to tag via key value applications running on a hadoop cluster.
Unfortunately, yarn.tags (supported by most frameworks) are a list of string.
This change adds the ability to have special yarn.tags containing a ':' to split key and value.

All those tags will appear in the ApplicationEvent under field user_tags which is a map of
string to string.

It will be exported as nested object on ElasticSearch or Parquet.