Spark storage status listener (#157)
Add a new Spark listener that focus on maintaining the storage state of the application as it is made on the SparkUI during runtime,
and firing events to visualise offline the storage evolution as timeseries

Storage is processed per RDD and per Executor. RDD storage keeps track of heap, off-heap and disk memory. Executor storage keeps track of heap, off-heap and disk memory for rdd blocks, broadcast blocks and stream blocks.

Events are fired a tasks end and when RDD is unpersisted. Two new events are created: RDDStorageStatus and ExecutorStorageStatus