Fix export of JvmStatistics event to HDFS (#221)
* Fix export of JvmStatistics event to HDFS

JvmStatistics uses sections identified by names and those sections contains properties identified by fields. This schema is not really enforced and method toProto does not work for embedded objects.

This commit adds the possibility to transform a GarmadonMessage body before converting it to proto. This way we can export only the section we want to a well defined proto.

Co-authored-by: William Montaz <w.montaz@criteo.com>